backend/requirements.txt
fastapi
uvicorn
python-pptx
python-multipart
openai
python-dotenv